feat!: Add support for credit cancelling (#385)
This commit adds support for cancelling ecocredits. An account holder can cancel credits that they own from any credit batch. Cancelled credits are removed from the holders tradeable balance and from the tradeable supply. Cancelling credits has no effect on the retirement balances and supplies. After cancellation it is as if the credit was never issued, except that it is tracked by amount_cancelled in the BatchInfo for the batch. This allows us to reconstruct the original amount issued from the current state.

Closes: #385